Feature/crud data kegiatan subkegiatan penugasan

+++ app/Http/Controllers/SubKegiatanController.php
@@ -53,11 +53,52 @@ class SubKegiatanController extends Controller
         ]);
     }
 
-    public function update(){}
-    public function destroy(Kegiatan $kegiatan, SubKegiatan $subKegiatan) {
+    public function update(Request $request, Kegiatan $kegiatan, SubKegiatan $subKegiatan){
+        // dd($request->all());
+
+        $validated = $request->validate([
+            'nama_sub_kegiatan' => ['required', 'string', 'max:255'],
+            'jenis_kegiatan' => ['required', 'string', 'max:255'],
+            'satuan_target' => ['required', 'string', 'max:255'],
+            'tanggal_mulai' => ['required', 'date', 'date_format:Y-m-d'],
+            'tanggal_selesai' => ['required',
+                                    'date',
+                                    'date_format:Y-m-d',
+                                    'after_or_equal:tanggal_mulai'],
+            'status' => ['required', 'in:Belum Mulai,Berjalan,Selesai'],
+        ]);
+
+        // 🔒 VALIDASI KEPEMILIKAN (PENTING!)
+        if ($subKegiatan->id_kegiatan !== $kegiatan->id_kegiatan) {
+            abort(403, 'Sub kegiatan tidak milik kegiatan ini');
+        }
+
+
+        try {
+            $subKegiatan->update($validated);
+            // Redirect dengan flash message
+            return redirect()
+                ->route('kegiatan.index', [
+                    'bidang' => $kegiatan->bidang->slug
+                ])
+                ->with('success', 'Sub Kegiatan berhasil diperbarui.');
+        } catch (\Exception $e) {
+            // dd($e->getMessage());
+            return redirect()->back()
+                ->with('error', 'Gagal memperbarui Sub Kegiatan. Silakan coba lagi.')
+                ->withInput();
+        }
+    }
+
+    public function delete(Kegiatan $kegiatan, SubKegiatan $subKegiatan) {
+        // optional safety check
+        if ($subKegiatan->id_kegiatan != $kegiatan->id_kegiatan) {
+            abort(403, 'Sub kegiatan tidak sesuai dengan kegiatan');
+        }
+
         $subKegiatan->delete();
 
-        return back()->with('success', 'Sub kegiatan berhasil dihapus');
+        return redirect()->back()->with('success', 'Sub kegiatan berhasil dihapus');
     }
 
 }
